Migrate search upsell page to unified AdminHeader (#47332)
* Migrate search upsell page header to unified AdminHeader via title prop.


* Add changelog entry for search package.


* Migrate search dashboard page header to AdminPage with unified AdminHeader.


* Delete unused search upsell header component and styles.


* update logo locator for e2e

---------
