+++ projects/packages/search/changelog/update-normalize-search-admin-header
@@ -1,0 +1,4 @@
+Significance: patch
+Type: changed
+
+Migrate upsell page header to use unified AdminHeader component.
