+++ projects/packages/search/src/dashboard/components/pages/dashboard-page.jsx
@@ -1,6 +1,6 @@
 import {
+	AdminPage,
 	JetpackFooter,
-	JetpackSearchLogo,
 	Button,
 	Container,
 	Col,
@@ -115,60 +115,73 @@ export default function DashboardPage( { isLoading = false } ) {
 			{ isPageLoading && <Loading /> }
 			{ ! isPageLoading && (
 				<div className="jp-search-dashboard-page">
-					<Header
-						isUpgradable={ ( isNewPricing && isFreePlan ) || ! supportsInstantSearch }
-						sendPaidPlanToCart={ sendPaidPlanToCart }
-					/>
-					{ hasConnectionError && (
-						<Container horizontalSpacing={ 3 } horizontalGap={ 3 }>
-							<Col lg={ 12 } md={ 12 } sm={ 12 }>
-								<ConnectionError />
-							</Col>
-						</Container>
-					) }
-					<MockedSearchInterface
-						supportsInstantSearch={ supportsInstantSearch }
-						supportsOnlyClassicSearch={ supportsOnlyClassicSearch }
-					/>
-					{ isNewPricing && supportsInstantSearch && (
-						<PlanInfo
-							hasIndex={ postCount !== 0 }
-							recordMeterInfo={ recordMeterInfo }
-							isFreePlan={ isFreePlan }
-							sendPaidPlanToCart={ sendPaidPlanToCart }
-						/>
-					) }
-					{ ! isNewPricing && supportsInstantSearch && (
-						<RecordMeter
-							postCount={ postCount }
-							postTypeBreakdown={ postTypeBreakdown }
-							tierMaximumRecords={ tierMaximumRecords }
-							lastIndexedDate={ lastIndexedDate }
-							postTypes={ postTypes }
-						/>
-					) }
-					<div className="jp-search-dashboard-bottom">
-						<ModuleControl
-							siteAdminUrl={ siteAdminUrl }
-							updateOptions={ updateOptions }
-							domain={ domain }
-							isDisabledFromOverLimit={ isOverLimit }
-							isInstantSearchPromotionActive={ isInstantSearchPromotionActive }
-							supportsOnlyClassicSearch={ supportsOnlyClassicSearch }
-							supportsSearch={ supportsSearch }
+					<AdminPage
+						title={ __( 'Search', 'jetpack-search-pkg' ) }
+						subTitle={ __(
+							'Help your visitors find exactly what they are looking for.',
+							'jetpack-search-pkg'
+						) }
+						actions={
+							( ( isNewPricing && isFreePlan ) || ! supportsInstantSearch ) && (
+								<Button variant="link" onClick={ sendPaidPlanToCart }>
+									{ __( 'Upgrade Jetpack Search', 'jetpack-search-pkg' ) }
+								</Button>
+							)
+						}
+						className="uses-new-admin-ui"
+						showFooter={ false }
+					>
+						{ hasConnectionError && (
+							<Container horizontalSpacing={ 3 } horizontalGap={ 3 }>
+								<Col lg={ 12 } md={ 12 } sm={ 12 }>
+									<ConnectionError />
+								</Col>
+							</Container>
+						) }
+						<MockedSearchInterface
 							supportsInstantSearch={ supportsInstantSearch }
-							isModuleEnabled={ isModuleEnabled }
-							isInstantSearchEnabled={ isInstantSearchEnabled }
-							isSavingEitherOption={ isSavingEitherOption }
-							isTogglingModule={ isTogglingModule }
-							isTogglingInstantSearch={ isTogglingInstantSearch }
+							supportsOnlyClassicSearch={ supportsOnlyClassicSearch }
 						/>
-					</div>
-					<Footer />
-					<NoticesList
-						notices={ notices }
-						handleLocalNoticeDismissClick={ handleLocalNoticeDismissClick }
-					/>
+						{ isNewPricing && supportsInstantSearch && (
+							<PlanInfo
+								hasIndex={ postCount !== 0 }
+								recordMeterInfo={ recordMeterInfo }
+								isFreePlan={ isFreePlan }
+								sendPaidPlanToCart={ sendPaidPlanToCart }
+							/>
+						) }
+						{ ! isNewPricing && supportsInstantSearch && (
+							<RecordMeter
+								postCount={ postCount }
+								postTypeBreakdown={ postTypeBreakdown }
+								tierMaximumRecords={ tierMaximumRecords }
+								lastIndexedDate={ lastIndexedDate }
+								postTypes={ postTypes }
+							/>
+						) }
+						<div className="jp-search-dashboard-bottom">
+							<ModuleControl
+								siteAdminUrl={ siteAdminUrl }
+								updateOptions={ updateOptions }
+								domain={ domain }
+								isDisabledFromOverLimit={ isOverLimit }
+								isInstantSearchPromotionActive={ isInstantSearchPromotionActive }
+								supportsOnlyClassicSearch={ supportsOnlyClassicSearch }
+								supportsSearch={ supportsSearch }
+								supportsInstantSearch={ supportsInstantSearch }
+								isModuleEnabled={ isModuleEnabled }
+								isInstantSearchEnabled={ isInstantSearchEnabled }
+								isSavingEitherOption={ isSavingEitherOption }
+								isTogglingModule={ isTogglingModule }
+								isTogglingInstantSearch={ isTogglingInstantSearch }
+							/>
+						</div>
+						<Footer />
+						<NoticesList
+							notices={ notices }
+							handleLocalNoticeDismissClick={ handleLocalNoticeDismissClick }
+						/>
+					</AdminPage>
 				</div>
 			) }
 		</>
@@ -248,27 +261,6 @@ const Footer = () => {
 					className="lg-col-span-12 md-col-span-8 sm-col-span-4"
 					useInternalLinks={ shouldUseInternalLinks() }
 				/>
-			</div>
-		</div>
-	);
-};
-
-const Header = ( { isUpgradable, sendPaidPlanToCart } ) => {
-	const buttonLinkArgs = {
-		children: __( 'Upgrade Jetpack Search', 'jetpack-search-pkg' ),
-		variant: 'link',
-		onClick: sendPaidPlanToCart,
-	};
-
-	return (
-		<div className="jp-search-dashboard-header jp-search-dashboard-wrap">
-			<div className="jp-search-dashboard-row">
-				<div className="lg-col-span-12 md-col-span-8 sm-col-span-4">
-					<div className="jp-search-dashboard-header__logo-container">
-						<JetpackSearchLogo className="jp-search-dashboard-header__masthead" />
-						{ isUpgradable && <Button { ...buttonLinkArgs } /> }
-					</div>
-				</div>
 			</div>
 		</div>
 	);
