+++ projects/packages/search/src/dashboard/components/pages/dashboard-page.scss
@@ -21,7 +21,7 @@ body {
 		sans-serif;
 }
 
-#wpcontent {
+#wpcontent:not:has(.uses-new-admin-ui) {
 	padding-left: 0 !important;
 }
 
@@ -79,17 +79,10 @@ body {
 		background-color: variables.$white;
 	}
 
-	.jp-search-dashboard-header,
 	.jp-search-dashboard-footer {
 		padding: 2.5em 0;
 
 		background-color: variables.$white;
-	}
-
-	.jp-search-dashboard-header__logo-container {
-		display: flex;
-		align-items: center;
-		justify-content: space-between;
 	}
 }
