+++ projects/packages/search/src/dashboard/components/pages/upsell-page/styles.scss
@@ -1,30 +1,5 @@
 @use "scss/variables";
 
-.jp-search-dashboard-upsell-page {
-	line-height: 1.5;
-	width: 100%;
-
-	h1 {
-		line-height: 1.2;
-	}
-
-	// Targets AdminPage component, which uses obfuscated CSS names.
-	> div {
-		margin-left: 0;
-	}
-
-	&__header {
-		display: flex;
-		justify-content: space-between;
-		flex-wrap: wrap;
-		gap: calc(var(--horizontal-spacing) * 3);
-	}
-
-	&__logo {
-		flex-shrink: 0;
-	}
-}
-
 .price-tip {
 	display: flex;
 	height: 20px;
